python/requirements.txt
kafka-python-ng
confluent_kafka
httpx